Create README.md
The process of taking action to increase organic traffic from search engines is known as search engine optimization, or SEO, marketing. To put it simply, organic traffic is the reverse of paid traffic; that is, the people who visit your content do not come from any sponsored promotion. Instead, the website design, the user-focused content on the front end, and the backend structure of the material are all credited with drawing in readers.